requirements.txt
PyQt5~=5.15.9
easyocr~=1.7.1
PyQt-Fluent-Widgets~=1.3.7
pyttsx3~=2.90
websocket-client~=1.6.4
opencv-python~=4.8.1.78
pyperclip~=1.8.2